requirements.txt
google-cloud-storage
Wand
